Create a dedicated component for the Save button and use it in the SettingPageTemplate;
Improve the NavigationGuard test coverage by checking the value of
'window.onbeforeunload' which is changed when the component unmounts.
