+++ Source/Plugins/Core/com.equella.core/js/__tests__/tsrc/components/NavigationGuard.test.tsx
@@ -1,10 +1,8 @@
 import * as React from "react";
 import { mount } from "enzyme";
-import * as module from "../../../tsrc/components/NavigationGuard";
+import * as NavigationGuardModule from "../../../tsrc/components/NavigationGuard";
 import { Router } from "react-router";
-import * as historyModule from "history";
-import { act } from "react-dom/test-utils";
-import ConfirmDialog from "../../../tsrc/components/ConfirmDialog";
+import { createMemoryHistory } from "history";
 
 /**
  * Need to mock the react-router context and function 'push'.
@@ -15,31 +13,29 @@ jest.mock("react-router-dom", () => ({
   }),
 }));
 
-describe("<NavigationGuard />", () => {
-  const history = historyModule.createMemoryHistory();
-  const mockBlockNavigationWithBrowser = jest.spyOn(
-    module,
-    "blockNavigationWithBrowser"
-  );
-  const component = mount(
-    <Router history={history}>
-      <module.NavigationGuard when />
-    </Router>
-  );
+const mockBlockNavigationWithBrowser = jest.spyOn(
+  NavigationGuardModule,
+  "blockNavigationWithBrowser"
+);
 
-  it("should call blockNavigationWithBrowser once in the useEffect", () => {
-    expect(component.find(ConfirmDialog)).toHaveLength(1);
+describe("<NavigationGuard />", () => {
+  const renderComponent = (when: boolean) => {
+    const history = createMemoryHistory();
+    return mount(
+      <Router history={history}>
+        <NavigationGuardModule.NavigationGuard when={when} />
+      </Router>
+    );
+  };
+
+  it("should call blockNavigationWithBrowser once when the component first mounts", () => {
+    renderComponent(true);
     expect(mockBlockNavigationWithBrowser).toHaveBeenCalledTimes(1);
   });
 
-  it("should show ConfirmDialog when route is going to change", () => {
-    // The dialog is closed.
-    expect(component.exists("h2")).not.toBe(true);
-
-    // The dialog is open when route change event is triggered.
-    history.listen(() => {
-      expect(component.find("h2").text()).toBe("Close without saving?");
-    });
-    act(() => history.push("home.do"));
+  it("should set 'window.onbeforeunload' to null when the component unmounts", () => {
+    const component = renderComponent(true);
+    component.unmount();
+    expect(window.onbeforeunload).toBeNull();
   });
 });
